prototype/requirements.txt
anthropic>=0.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
websockets>=12.0
python-multipart>=0.0.6
numpy>=1.26.0
